requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.4.2
idna==2.8
ipaddress==1.0.22
lxml==3.8.0
pycparser==2.19
pyOpenSSL==18.0.0
six==1.12.0
xmlsig==0.1.1
